requirements.txt
keyboard == 0.13
pillow == 10.2
pytest == 8

